+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add store users
 - Add store languages
 - Add store channels
+- Add fixed header admin
 ### Changed
 - Updated page store with action returning a promise
 - Removed container class from basic template index